requirements.txt
frappe
erpnext
python-barcode
pillow
